Add basic game loop
The basic game loop is only designed for a board with single
recursion depth. When the left mouse button is clicked, the view
calculates which column was clicked and the board is updated in
the lowest free spot. The player alternates after a move. If the
board is completed then nothing happens on mouse click.